# labels/cidr: Cache GetCIDRLabels computation