@joamaki 2023-09-01
![image](https://github.com/cilium/cilium/assets/1102396/f2b6797c-93e4-45e0-b1f5-873125426812)

This significantly reduces allocation rate in ops, but seems to increase memory usage and allocation rate in bytes/s. It might be that the `sync.Pool` based memoization has poor hit rate. I'll look into a LRU based approach instead and see if that performs better. Based on the graph I don't feel comfortable with the trade-off.